+++ pyproject.toml
@@ -297,8 +297,8 @@ betty = 'betty.console:main_standalone'
 'witness' = 'betty.roles.witness:Witness'
 
 [project.entry-points.'betty.http_rate_limit']
-'wikipedia-action-api' = 'betty.plugins.rate_limit.wikipedia_action_api:WIKIPEDIA_ACTION_API'
-'wikipedia-rest-api' = 'betty.plugins.rate_limit.wikipedia_rest_api:WIKIPEDIA_REST_API'
+'wikipedia-action-api' = 'betty.http_client.rate_limits.wikipedia_action_api:WIKIPEDIA_ACTION_API'
+'wikipedia-rest-api' = 'betty.http_client.rate_limits.wikipedia_rest_api:WIKIPEDIA_REST_API'
 
 [project.entry-points.'betty.serializer']
 'json' = 'betty.serializers.json:Json'
